Test/product integration

+++ src/main/java/com/groom/e_commerce/product/domain/entity/ProductVariant.java
@@ -57,7 +57,7 @@ public class ProductVariant extends BaseEntity {
 	private String optionName;
 
 	@Column(name = "price", nullable = false, precision = 12, scale = 2)
-	private BigDecimal price;
+	private Long price;
 
 	@Column(name = "stock_quantity", nullable = false)
 	private Integer stockQuantity;
@@ -68,7 +68,7 @@ public class ProductVariant extends BaseEntity {
 
 	@Builder
 	public ProductVariant(Product product, String skuCode, List<UUID> optionValueIds,
-		String optionName, BigDecimal price, Integer stockQuantity) {
+		String optionName, Long price, Integer stockQuantity) {
 		this.product = product;
 		this.skuCode = skuCode;
 		this.optionValueIds = optionValueIds;
@@ -78,7 +78,7 @@ public class ProductVariant extends BaseEntity {
 		this.status = VariantStatus.ON_SALE;
 	}
 
-	public void update(String optionName, BigDecimal price, Integer stockQuantity) {
+	public void update(String optionName, Long price, Integer stockQuantity) {
 		if (optionName != null) {
 			this.optionName = optionName;
 		}
